fixed bug when parsing R custom functions into PMML
To see an example of the type of code that the bug fixed, look at the new test
added in the test-r-to-pmml-custom-functions file. Code that was parsing and
converting these data frame expressions had a space before the closing tag of
the TableLocator node. However, the code that was replacing the TableLocator
node with one where the children had a call to the inner function, did not have
the space. The code change in the r-to-pmml-define-functin.R adds this space.

However, this broke the test "When multiple expressions are present within a
function call..." in the test-r-to-pmml-df-code.R file. The test was failing
because the code that was adding the TableLocator node for the test code did
not have a space at the end, so the change mentioned above was resulting in the
added TableLocator node not being replaced with its inner function equivalent.
The fix in the r-to-pmml-data-frame.R file fixes this by adding the space.