fix issue with parsing non-wildcard data frame code within custom functions
for an example of the code that was failing look at the newly added test
in test-r-to-pmml-custom-functions. This was happening because of a
missing space before the closing space of the TableLocator node. The
missing space was causing the code that replaces the TableLocator node
with the correct version that uses an inner function name to not find
the node.

This change resulted in the test "Non-wildcard column access expressions inside
functions throw an error" to fail (note: this test title is wrong)
because now the code was replacing all applicable TableLocator, even
those that were not within the scope of the function. The code change
also added a check to ensure only tables that were within the scope of
the function were replaced.