fixed bug with parsing data frame expressions within function calls
for an example of the code that was causing the issue take a look at the
new test in the test-r-to-pmml-custom-functions file.

The issue was that the TableLocator node had a location value of
"Taxonomy" instead of "local" when the table variable was within the
function scope. This was because the list of function scope variables
was being hardcoded to an empty string. The fix was to add in a new
parameter that had the scope variables to the function. Updated all
function calls.